requirements.txt
# Data Manipulation
pandas
numpy
polars

# Visualization
matplotlib
seaborn
plotly
wordcloud
missingno

# Statistical Analysis
scipy
statsmodels
prophet

# NLP Libraries
nltk
textstat
langdetect

# Geospatial
folium
shapely
h3

# Utilities
tqdm
python-dotenv
pyyaml
requests
sqlalchemy
python-jose
passlib[bcrypt]
chromadb
lime

# Jupyter
jupyter
ipywidgets

# Web Application & NLP
streamlit
scikit-learn
networkx
fastapi
uvicorn
